worker-service/requirements.txt
SQLAlchemy
psycopg2-binary
pika
minio
Pillow

